restructure executor tests
